requirements.txt
fastapi
uvicorn
openai
langchain
faiss-cpu
PyMuPDF
python-dotenv

sentence-transformers==2.4.0
transformers==4.35.2
huggingface-hub==0.20.3